+++ services/storagecache/mgmt/2020-03-01/storagecache/caches.go
@@ -220,13 +220,12 @@ func (client CachesClient) DeleteSender(req *http.Request) (future CachesDeleteF
 
 // DeleteResponder handles the response to the Delete request. The method always
 // closes the http.Response Body.
-func (client CachesClient) DeleteResponder(resp *http.Response) (result SetObject, err error) {
+func (client CachesClient) DeleteResponder(resp *http.Response) (result autorest.Response, err error) {
 	err = autorest.Respond(
 		resp,
 		azure.WithErrorUnlessStatusCode(http.StatusOK, http.StatusAccepted, http.StatusNoContent),
-		autorest.ByUnmarshallingJSON(&result),
 		autorest.ByClosing())
-	result.Response = autorest.Response{Response: resp}
+	result.Response = resp
 	return
 }
 
@@ -306,13 +305,12 @@ func (client CachesClient) FlushSender(req *http.Request) (future CachesFlushFut
 
 // FlushResponder handles the response to the Flush request. The method always
 // closes the http.Response Body.
-func (client CachesClient) FlushResponder(resp *http.Response) (result SetObject, err error) {
+func (client CachesClient) FlushResponder(resp *http.Response) (result autorest.Response, err error) {
 	err = autorest.Respond(
 		resp,
 		azure.WithErrorUnlessStatusCode(http.StatusOK, http.StatusAccepted, http.StatusNoContent),
-		autorest.ByUnmarshallingJSON(&result),
 		autorest.ByClosing())
-	result.Response = autorest.Response{Response: resp}
+	result.Response = resp
 	return
 }
 
@@ -703,13 +701,12 @@ func (client CachesClient) StartSender(req *http.Request) (future CachesStartFut
 
 // StartResponder handles the response to the Start request. The method always
 // closes the http.Response Body.
-func (client CachesClient) StartResponder(resp *http.Response) (result SetObject, err error) {
+func (client CachesClient) StartResponder(resp *http.Response) (result autorest.Response, err error) {
 	err = autorest.Respond(
 		resp,
 		azure.WithErrorUnlessStatusCode(http.StatusOK, http.StatusAccepted, http.StatusNoContent),
-		autorest.ByUnmarshallingJSON(&result),
 		autorest.ByClosing())
-	result.Response = autorest.Response{Response: resp}
+	result.Response = resp
 	return
 }
 
@@ -788,13 +785,12 @@ func (client CachesClient) StopSender(req *http.Request) (future CachesStopFutur
 
 // StopResponder handles the response to the Stop request. The method always
 // closes the http.Response Body.
-func (client CachesClient) StopResponder(resp *http.Response) (result SetObject, err error) {
+func (client CachesClient) StopResponder(resp *http.Response) (result autorest.Response, err error) {
 	err = autorest.Respond(
 		resp,
 		azure.WithErrorUnlessStatusCode(http.StatusOK, http.StatusAccepted, http.StatusNoContent),
-		autorest.ByUnmarshallingJSON(&result),
 		autorest.ByClosing())
-	result.Response = autorest.Response{Response: resp}
+	result.Response = resp
 	return
 }
 
@@ -967,12 +963,11 @@ func (client CachesClient) UpgradeFirmwareSender(req *http.Request) (future Cach
 
 // UpgradeFirmwareResponder handles the response to the UpgradeFirmware request. The method always
 // closes the http.Response Body.
-func (client CachesClient) UpgradeFirmwareResponder(resp *http.Response) (result SetObject, err error) {
+func (client CachesClient) UpgradeFirmwareResponder(resp *http.Response) (result autorest.Response, err error) {
 	err = autorest.Respond(
 		resp,
 		azure.WithErrorUnlessStatusCode(http.StatusOK, http.StatusCreated, http.StatusAccepted, http.StatusNoContent),
-		autorest.ByUnmarshallingJSON(&result),
 		autorest.ByClosing())
-	result.Response = autorest.Response{Response: resp}
+	result.Response = resp
 	return
 }